requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
bleach==5.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.4
django-appconf==1.0.5
django-compressor==4.3
django-compressor-toolkit==0.6.0
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
Markdown==3.3.4
martor==1.6.15
multidict==6.0.4
numpy==1.24.1
pandas==1.5.2
python-dateutil==2.8.2
python-nmap==0.7.1
pytz==2022.7
rcssmin==1.1.1
requests==2.28.1
rjsmin==1.2.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.13
webencodings==0.5.1
yarl==1.8.2
